refactor: assertLocaleData returns a value
