+++ src/internal/locale-proxy.ts
@@ -1,5 +1,6 @@
 import type { LocaleDefinition } from '../definitions';
 import { FakerError } from '../errors/faker-error';
+import { assertLocaleData } from './assert-locale-data';
 
 /**
  * A proxy for LocaleDefinition that marks all properties as required and throws an error when an entry is accessed that is not defined.
@@ -55,31 +56,6 @@ export function createLocaleProxy(locale: LocaleDefinition): LocaleProxy {
 }
 
 /**
- * Checks that the value is not null or undefined and throws an error if it is.
- *
- * @param value The value to check.
- * @param path The path to the locale data.
- */
-export function assertLocaleData<T>(
-  value: T,
-  ...path: string[]
-): asserts value is NonNullable<T> {
-  if (value === null) {
-    throw new FakerError(
-      `The locale data for '${path.join('.')}' aren't applicable to this locale.
-  If you think this is a bug, please report it at: https://github.com/faker-js/faker`
-    );
-  } else if (value === undefined) {
-    throw new FakerError(
-      `The locale data for '${path.join('.')}' are missing in this locale.
-  If this is a custom Faker instance, please make sure all required locales are used e.g. '[de_AT, de, en, base]'.
-  Please contribute the missing data to the project or use a locale/Faker instance that has these data.
-  For more information see https://fakerjs.dev/guide/localization.html`
-    );
-  }
-}
-
-/**
  * Creates a proxy for a category that throws an error when accessing an undefined property.
  *
  * @param categoryName The name of the category.
@@ -106,8 +82,7 @@ function createCategoryProxy<
         return value;
       }
 
-      assertLocaleData(value, categoryName, entryName.toString());
-      return value;
+      return assertLocaleData(value, categoryName, entryName.toString());
     },
 
     set: throwReadOnlyError,
