+++ src/modules/date/index.ts
@@ -1,8 +1,8 @@
 import type { DateEntryDefinition } from '../../definitions';
 import { FakerError } from '../../errors/faker-error';
 import type { Faker } from '../../faker';
+import { assertLocaleData } from '../../internal/assert-locale-data';
 import { toDate } from '../../internal/date';
-import { assertLocaleData } from '../../internal/locale-proxy';
 import { SimpleModuleBase } from '../../internal/module-base';
 
 /**
@@ -657,9 +657,9 @@ export class DateModule extends SimpleDateModule {
       type = useContext ? 'wide_context' : 'wide';
     }
 
-    const values = source[type];
-    assertLocaleData(values, 'date.month', type);
-    return this.faker.helpers.arrayElement(values);
+    return this.faker.helpers.arrayElement(
+      assertLocaleData(source[type], 'date.month', type)
+    );
   }
 
   /**
@@ -709,9 +709,9 @@ export class DateModule extends SimpleDateModule {
       type = useContext ? 'wide_context' : 'wide';
     }
 
-    const values = source[type];
-    assertLocaleData(values, 'date.weekday', type);
-    return this.faker.helpers.arrayElement(values);
+    return this.faker.helpers.arrayElement(
+      assertLocaleData(source[type], 'date.weekday', type)
+    );
   }
 
   /**
